+++ utils/src/com/cloud/utils/encoding/URLEncoder.java
@@ -71,7 +71,7 @@ public class URLEncoder {
         }
 
         for (int i = 0; i < path.length(); i++) {
-            int c = (int)path.charAt(i);
+            int c = path.charAt(i);
             // NOTICE - !isPureAscii(path.charAt(i)) check was added by
             // CloudStack
             if (safeCharacters.get(c) || !isPureAscii(path.charAt(i))) {
@@ -90,8 +90,8 @@ public class URLEncoder {
                     // Converting each byte in the buffer
                     byte toEncode = ba[j];
                     rewrittenPath.append('%');
-                    int low = (int)(toEncode & 0x0f);
-                    int high = (int)((toEncode & 0xf0) >> 4);
+                    int low = toEncode & 0x0f;
+                    int high = (toEncode & 0xf0) >> 4;
                     rewrittenPath.append(hexadecimal[high]);
                     rewrittenPath.append(hexadecimal[low]);
                 }
